Moved  package to ch.stockmanager.server
