+++ src/main/java/ch/stockmanager/server/controllers/ProductController.java
@@ -8,10 +8,10 @@ import java.util.List;
 import io.javalin.http.Context;
 import io.javalin.http.HttpStatus;
 
-import ch.stockmanager.types.Order;
-import ch.stockmanager.types.ProductDateQuantity;
+import ch.stockmanager.server.types.Order;
+import ch.stockmanager.server.types.ProductDateQuantity;
 import ch.stockmanager.server.util.*;
-import ch.stockmanager.types.Product;
+import ch.stockmanager.server.types.Product;
 
 public class ProductController extends Controller {
 	private final DBInfo dbInfo;
