+++ src/main/java/ch/stockmanager/server/types/Lot.java
@@ -1,4 +1,4 @@
-package ch.stockmanager.types;
+package ch.stockmanager.server.types;
 
 import java.util.Date;
