+++ src/main/java/ch/stockmanager/server/types/ProductDateQuantity.java
@@ -1,4 +1,4 @@
-package ch.stockmanager.types;
+package ch.stockmanager.server.types;
 
 @SuppressWarnings("unused")
 public class ProductDateQuantity {
