+++ src/main/java/ch/stockmanager/server/types/ProductYearSegment.java
@@ -1,4 +1,4 @@
-package ch.stockmanager.types;
+package ch.stockmanager.server.types;
 
 public class ProductYearSegment {
 	public long productCode;
