+++ src/main/java/ch/stockmanager/server/types/YearSegment.java
@@ -1,4 +1,4 @@
-package ch.stockmanager.types;
+package ch.stockmanager.server.types;
 
 public class YearSegment {
 	public int id;
